Update ecl.h
Allocating more memory to malloc of evalc
Testing if files "tmp.c" and "tmp.so" are present before deletion
Including "unistd.h" header for both C and C++

+++ ecl.h
@@ -3,6 +3,7 @@
 #ifdef __cplusplus
 #include<cstring>
 #include<dlfcn.h>
+#include<unistd.h>
 #ifndef ECL_COMPILE_INSTRUCTION
 	#define ECL_COMPILE_INSTRUCTION "g++ -fPIC -shared -o tmp.so tmp.c"
 #endif
@@ -14,6 +15,7 @@
 #include<stdlib.h>
 #include<dlfcn.h>
 #include<string.h>
+#include<unistd.h>
 #ifndef ECL_COMPILE_INSTRUCTION
 	#define ECL_COMPILE_INSTRUCTION "gcc -fPIC -shared -o tmp.so tmp.c" 
 #endif
@@ -27,20 +29,20 @@
 	lib=dlopen("./tmp.so",RTLD_LAZY);\
 	ECL_GET_DL_POINTER;\
 		dlclose(lib); init();\
-		system("rm tmp.so");
+		if(access("tmp.so",F_OK)==0) system("rm tmp.so");
 	
 void evalc(const char data[256]){
 	FILE *fptr;
 	fptr=fopen("tmp.c","w"); fprintf(fptr,"%s",data); fclose(fptr);
 		ECL_COMPILE; ECL_LOAD_LIB();
-		system("rm tmp.c");
+		if(access("tmp.c",F_OK)==0) system("rm tmp.c");
 }
 
 void evalf(const char *file){
-	char *buffer=(char*)malloc(256);
+	char *buffer=(char*)malloc(512);
 	strcat(buffer,"cp "); strcat(buffer,file); strcat(buffer," tmp.c "); system(buffer);
 		ECL_COMPILE; ECL_LOAD_LIB();
-		system("rm tmp.c");
+		if(access("tmp.c",F_OK)==0)  system("rm tmp.c");
 }
 
 #endif
